feat: 기존 api 관련 로직 server app으로 분리
