+++ apps/server/src/data/db.json
@@ -1,6 +1,22 @@
 {
   "posts": [
     {
+      "id": "1738902606053",
+      "title": "test",
+      "content": "test",
+      "author": "test",
+      "createdAt": "2025-02-07T04:30:06.053Z",
+      "updatedAt": "2025-02-07T04:30:06.053Z"
+    },
+    {
+      "id": "1738901411140",
+      "title": "123123",
+      "content": "123123",
+      "author": "123",
+      "createdAt": "2025-02-07T04:10:11.140Z",
+      "updatedAt": "2025-02-07T04:10:11.140Z"
+    },
+    {
       "id": "17358013402371",
       "title": "개발자 커리어 고민",
       "content": "프론트엔드와 백엔드 사이에서 고민이 많습니다.\n어떤 길로 가야 할지 조언 부탁드려요.",
